requirements.txt
cryptography==42.0.8
Eel==0.16.0
minecraft_launcher_lib==6.5
psutil==5.9.8
PyMsgBox==1.0.9
pywebview==5.1
Requests==2.32.3
show_in_file_manager==1.1.5
git+https://github.com/bottlepy/bottle.git